+++ test/caching_test.rb
@@ -1,4 +1,5 @@
 require File.expand_path('../support/test_helper', __FILE__)
+require 'tempfile'
 
 class CachingTestTest < Minitest::Test
   def setup