+++ test/fragment_resolution_test.rb
@@ -19,11 +19,11 @@ class FragmentResolutionTest < Minitest::Test
     refute_valid schema, data
     assert_valid schema, data, :fragment => "#/properties/a"
 
-    assert_raises JSON::Schema::SchemaError do
+    assert_raises JSON::Schema::Pointer::PointerSyntaxError do
       JSON::Validator.validate!(schema,data,:fragment => "/properties/a")
     end
 
-    assert_raises JSON::Schema::SchemaError do
+    assert_raises JSON::Schema::Pointer::ReferenceError do
       JSON::Validator.validate!(schema,data,:fragment => "#/properties/b")
     end
   end
@@ -60,6 +60,34 @@ class FragmentResolutionTest < Minitest::Test
     refute_valid schema, {}, :fragment => "#/foo/bar"
   end
 
+  def test_fragment_resolution_with_array
+    document = {
+      "definitions" => {
+        "schemas" => [
+          {
+            "type" => "object",
+            "required" => ["a"],
+            "properties" => {
+              "a" => {
+                "type" => "object",
+              }
+            }
+          },
+          {
+            "type" => "object",
+            "properties" => {
+              "b" => {"type" => "integer" }
+            }
+          }
+        ]
+      }
+    }
+
+    data = {"b" => 5}
+    refute_valid document, data, :fragment => "#/definitions/schemas/0"
+    assert_valid document, data, :fragment => "#/definitions/schemas/1"
+  end
+
   def test_array_fragment_resolution
     schema = {
       "type" => "object",
@@ -79,5 +107,32 @@ class FragmentResolutionTest < Minitest::Test
 
     assert_valid schema, 5, :fragment => "#/properties/a/anyOf/0"
     refute_valid schema, 5, :fragment => "#/properties/a/anyOf/1"
+  end
+  def test_fragment_resolution_with_special_chars
+    document = {
+      "de~fi/nitions" => {
+        "schemas" => [
+          {
+            "type" => "object",
+            "required" => ["a"],
+            "properties" => {
+              "a" => {
+                "type" => "object",
+              }
+            }
+          },
+          {
+            "type" => "object",
+            "properties" => {
+              "b" => {"type" => "integer" }
+            }
+          }
+        ]
+      }
+    }
+
+    data = {"b" => 5}
+    refute_valid document, data, :fragment => "#/de~0fi~1nitions/schemas/0"
+    assert_valid document, data, :fragment => "#/de~0fi~1nitions/schemas/1"
   end
 end