+++ CHANGELOG.md
@@ -16,6 +16,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Rescue URI error when initializing a data string that contains a colon
 - Fragments with an odd number of components no longer raise an `undefined method `validate'`
   error
+- Incorrect instantiation of json documents as schemas when resolving fragments
+- Validation of a schema with validate_schema and a fragment validates the schema at the fragment pointer
+- Handling of fragment strings containing URI escaped values and pointers escaped with ~0 / ~1
 
 ## [2.8.0] - 2017-02-07